requirements.txt
opencv-python==4.9.0.80
opencv-contrib-python==4.9.0.80
numpy>=1.26.0
requests==2.31.0
Pillow==10.0.1
